+++ tests/test_models/conftest.py
@@ -1,3 +1,4 @@
+import numpy as np
 import pytest
 
 from etna.datasets import generate_ar_df
@@ -16,3 +17,15 @@ def new_format_exog():
     exog = generate_ar_df(periods=60, start_time="2021-06-01", n_segments=2)
     df = TSDataset.to_dataset(exog)
     return df
+
+
+@pytest.fixture()
+def dfs_w_exog():
+    df = generate_ar_df(start_time="2021-01-01", periods=105, n_segments=1)
+    df["f1"] = np.sin(df["target"])
+    df["f2"] = np.cos(df["target"])
+
+    df.drop(columns=["segment"], inplace=True)
+    train = df.iloc[:-5]
+    test = df.iloc[-5:]
+    return train, test
